Remove AnimationFrameVisualizer, deprecate old atframe methods.
